Master

+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/web/controller/UserManagementController.java
@@ -2,30 +2,38 @@ package gov.healthit.chpl.web.controller;
 
 
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
-import java.util.Properties;
 import java.util.Set;
 
-import javax.mail.Message;
 import javax.mail.MessagingException;
-import javax.mail.PasswordAuthentication;
-import javax.mail.Session;
-import javax.mail.Transport;
 import javax.mail.internet.AddressException;
-import javax.mail.internet.InternetAddress;
-import javax.mail.internet.MimeMessage;
 
+import org.apache.log4j.LogManager;
+import org.apache.log4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.acls.domain.BasePermission;
+import org.springframework.security.acls.domain.PrincipalSid;
+import org.springframework.util.StringUtils;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+
+import gov.healthit.chpl.auth.AuthPropertiesConsumer;
+import gov.healthit.chpl.auth.SendMailUtil;
 import gov.healthit.chpl.auth.Util;
 import gov.healthit.chpl.auth.authentication.Authenticator;
-import gov.healthit.chpl.auth.authentication.LoginCredentials;
 import gov.healthit.chpl.auth.dto.InvitationDTO;
 import gov.healthit.chpl.auth.dto.UserDTO;
 import gov.healthit.chpl.auth.dto.UserPermissionDTO;
 import gov.healthit.chpl.auth.json.GrantRoleJSONObject;
 import gov.healthit.chpl.auth.json.User;
-import gov.healthit.chpl.auth.json.UserCreationJSONObject;
-import gov.healthit.chpl.auth.json.UserCreationWithRolesJSONObject;
 import gov.healthit.chpl.auth.json.UserInfoJSONObject;
 import gov.healthit.chpl.auth.json.UserInvitation;
 import gov.healthit.chpl.auth.json.UserListJSONObject;
@@ -35,57 +43,84 @@ import gov.healthit.chpl.auth.permission.UserPermissionRetrievalException;
 import gov.healthit.chpl.auth.user.UserCreationException;
 import gov.healthit.chpl.auth.user.UserManagementException;
 import gov.healthit.chpl.auth.user.UserRetrievalException;
+import gov.healthit.chpl.dao.EntityCreationException;
 import gov.healthit.chpl.dao.EntityRetrievalException;
+import gov.healthit.chpl.domain.ActivityConcept;
 import gov.healthit.chpl.domain.AuthorizeCredentials;
-import gov.healthit.chpl.domain.CertificationBodyPermission;
 import gov.healthit.chpl.domain.CreateUserFromInvitationRequest;
 import gov.healthit.chpl.dto.CertificationBodyDTO;
+import gov.healthit.chpl.manager.ActivityManager;
 import gov.healthit.chpl.manager.CertificationBodyManager;
 import gov.healthit.chpl.manager.InvitationManager;
 
-import org.apache.log4j.LogManager;
-import org.apache.log4j.Logger;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.acls.domain.BasePermission;
-import org.springframework.security.acls.domain.PrincipalSid;
-import org.springframework.security.acls.model.Permission;
-import org.springframework.util.StringUtils;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
-
 @RestController
 @RequestMapping("/users")
-public class UserManagementController {
+public class UserManagementController extends AuthPropertiesConsumer {
 	
 	@Autowired UserManager userManager;
 	@Autowired CertificationBodyManager acbManager;
 	@Autowired InvitationManager invitationManager;
 	@Autowired private Authenticator authenticator;
+	@Autowired private ActivityManager activityManager;
 	
 	private static final Logger logger = LogManager.getLogger(UserManagementController.class);
-	
+    
 	@RequestMapping(value="/create", method= RequestMethod.POST, 
 			consumes= MediaType.APPLICATION_JSON_VALUE,
 			produces="application/json; charset=utf-8")
 	public User createUser(@RequestBody CreateUserFromInvitationRequest userInfo) 
-			throws InvalidArgumentsException, UserCreationException, UserRetrievalException, EntityRetrievalException {
-		boolean validHash = invitationManager.isHashValid(userInfo.getHash());
-		if(!validHash) {
-			throw new InvalidArgumentsException("Provided hash is not valid in the database. The hash is valid for up to 3 days from when it is assigned.");
-		}
+			throws InvalidArgumentsException, UserCreationException, UserRetrievalException, 
+			EntityRetrievalException, MessagingException, JsonProcessingException, EntityCreationException {
 		
 		if(userInfo.getUser() == null || userInfo.getUser().getSubjectName() == null) {
 			throw new InvalidArgumentsException("Username ('subject name') is required.");
 		}
 		
-		InvitationDTO invitation = invitationManager.getByHash(userInfo.getHash());
+		InvitationDTO invitation = invitationManager.getByInvitationHash(userInfo.getHash());
+		if(invitation == null || invitation.isExpired()) {
+			throw new InvalidArgumentsException("Provided hash is not valid in the database. The hash is valid for up to 3 days from when it is assigned.");
+		}
+		
 		UserDTO createdUser = invitationManager.createUserFromInvitation(invitation, userInfo.getUser());
+		
+		//get the invitation again to get the new hash
+		invitation = invitationManager.getById(invitation.getId());
+		
+		//send email for user to confirm email address	
+		String htmlMessage = "<p>Thank you for setting up your administrator account on ONC's Open Data CHPL. " +
+					"Please click the link below to activate your account: <br/>" +
+					"http://" + getProps().getProperty("chplServer") + "/#/registration/confirm-user/" + invitation.getConfirmToken() +
+				"</p>" +
+				"<p>If you have any questions, please contact Scott Purnell-Saunders at Scott.Purnell-Saunders@hhs.gov.</p>" +
+				"<p>The Open Data CHPL Team</p>";
+
+		SendMailUtil emailUtils = new SendMailUtil();
+		emailUtils.sendEmail(createdUser.getEmail(), "Confirm CHPL Administrator Account", htmlMessage);
+		
+		String activityDescription = "User "+createdUser.getSubjectName()+" was created.";
+		activityManager.addActivity(ActivityConcept.ACTIVITY_CONCEPT_USER, createdUser.getId(), activityDescription, null, createdUser, createdUser.getId());
+		
+		return new User(createdUser);
+	}
+	
+	@RequestMapping(value="/confirm", method= RequestMethod.POST, 
+			consumes= MediaType.APPLICATION_JSON_VALUE,
+			produces="application/json; charset=utf-8")
+	public User createUser(@RequestBody String hash) 
+			throws InvalidArgumentsException, UserRetrievalException, EntityRetrievalException, MessagingException, JsonProcessingException, EntityCreationException {
+		InvitationDTO invitation = invitationManager.getByConfirmationHash(hash);
+
+		if(invitation == null || invitation.isExpired())
+		{
+			throw new InvalidArgumentsException("Provided hash is not valid in the database. The hash is valid for up to 3 days from when it is assigned.");
+		}
+		
+		UserDTO createdUser = invitationManager.confirmAccountEmail(invitation);
+		
+		String activityDescription = "User "+createdUser.getSubjectName()+" was confirmed.";
+		activityManager.addActivity(ActivityConcept.ACTIVITY_CONCEPT_USER, createdUser.getId(), activityDescription, createdUser, createdUser, createdUser.getId());
+		
+		
 		return new User(createdUser);
 	}
 	
@@ -101,9 +136,8 @@ public class UserManagementController {
 				StringUtils.isEmpty(credentials.getPassword())) {
 			throw new InvalidArgumentsException("Username, Password, and Token are all required.");
 		}
-		
-		boolean validHash = invitationManager.isHashValid(credentials.getHash());
-		if(!validHash) {
+		InvitationDTO invitation = invitationManager.getByInvitationHash(credentials.getHash());
+		if(invitation == null || invitation.isExpired()) {
 			throw new InvalidArgumentsException("Provided hash is not valid in the database. The hash is valid for up to 3 days from when it is assigned.");
 		}
 		
@@ -112,7 +146,6 @@ public class UserManagementController {
 			throw new UserRetrievalException("The user " + credentials.getUserName() + " could not be authenticated.");
 		}
 		
-		InvitationDTO invitation = invitationManager.getByHash(credentials.getHash());
 		invitationManager.updateUserFromInvitation(invitation, userToUpdate);
 		
 		String jwt = authenticator.getJWT(credentials);
@@ -142,21 +175,39 @@ public class UserManagementController {
 			createdInvite = invitationManager.inviteWithAcbAccess(invitation.getEmailAddress(), invitation.getAcbId(), invitation.getPermissions());
 		}
 		
-		//send email
-		//sendEmail(createdInvite);
+		//send email		
+		String htmlMessage = "<p>Hi,</p>" +
+				"<p>You've been invited to be an Administrator on the ONC's Open Data CHPL, " +
+					"which will allow you to manage certified product listings on the CHPL. " +
+					"Please click the link below to create your account: <br/>" +
+					"http://" + getProps().getProperty("chplServer") + "/#/registration/create-user/"+ createdInvite.getInviteToken() +
+				"</p>" +
+				"<p>If you have any questions, please contact Scott Purnell-Saunders at Scott.Purnell-Saunders@hhs.gov.</p>" +
+				"<p>Take care,<br/> " +
+				 "The Open Data CHPL Team</p>";
+
+		SendMailUtil emailUtils = new SendMailUtil();
+		emailUtils.sendEmail(createdInvite.getEmail(), "Open Data CHPL Administrator Invitation", htmlMessage);
 		
 		UserInvitation result = new UserInvitation(createdInvite);
 		return result;
-	}
+	}	
 	
 	@RequestMapping(value="/update", method= RequestMethod.POST, 
 			consumes= MediaType.APPLICATION_JSON_VALUE,
 			produces="application/json; charset=utf-8")
-	public User updateUserDetails(@RequestBody User userInfo) throws UserRetrievalException, UserPermissionRetrievalException {
+	public User updateUserDetails(@RequestBody User userInfo) throws UserRetrievalException, UserPermissionRetrievalException, JsonProcessingException, EntityCreationException, EntityRetrievalException {
+		
 		if(userInfo.getUserId() <= 0) {
 			throw new UserRetrievalException("Cannot update user with ID less than 0");
 		}
+		
+		UserDTO before = userManager.getById(userInfo.getUserId());
 		UserDTO updated = userManager.update(userInfo);
+		
+		String activityDescription = "User "+userInfo.getSubjectName()+" was updated.";
+		activityManager.addActivity(ActivityConcept.ACTIVITY_CONCEPT_USER, before.getId(), activityDescription, before, updated);
+		
 		return new User(updated);
 	}
 	
@@ -164,11 +215,15 @@ public class UserManagementController {
 	@RequestMapping(value="/{userId}/delete", method= RequestMethod.POST,
 			produces="application/json; charset=utf-8")
 	public String deleteUser(@PathVariable("userId") Long userId) 
-			throws UserRetrievalException, UserManagementException, UserPermissionRetrievalException {
+			throws UserRetrievalException, UserManagementException, UserPermissionRetrievalException, JsonProcessingException, EntityCreationException, EntityRetrievalException {
 		if(userId <= 0) {
 			throw new UserRetrievalException("Cannot delete user with ID less than 0");
 		}
 		UserDTO toDelete = userManager.getById(userId);
+		
+		if(toDelete == null) {
+			throw new UserRetrievalException("Could not find user with id " + userId);
+		}
 		
 		//delete the acb permissions for that user
 		acbManager.deletePermissionsForUser(toDelete);
@@ -176,26 +231,19 @@ public class UserManagementController {
 		//delete the user
 		userManager.delete(toDelete);
 		
+		String activityDescription = "User "+toDelete.getSubjectName()+" was deleted.";
+		activityManager.addActivity(ActivityConcept.ACTIVITY_CONCEPT_USER, toDelete.getId(), activityDescription, toDelete, null);
+		
+		
 		return "{\"deletedUser\" : true }";
 	}
 	
-	
-	@RequestMapping(value="/reset_password", method= RequestMethod.POST, 
-			consumes= MediaType.APPLICATION_JSON_VALUE,
-			produces="application/json; charset=utf-8")
-	public String resetPassword(@RequestBody LoginCredentials newCredentials) throws UserRetrievalException {
-		
-		userManager.updateUserPassword(newCredentials.getUserName(), newCredentials.getPassword());
-		return "{\"passwordUpdated\" : true }";
-	
-	}	
-	
+
 	@RequestMapping(value="/grant_role", method= RequestMethod.POST, 
 			consumes= MediaType.APPLICATION_JSON_VALUE,
 			produces="application/json; charset=utf-8")
-	public String grantUserRole(@RequestBody GrantRoleJSONObject grantRoleObj) throws InvalidArgumentsException, UserRetrievalException, UserManagementException, UserPermissionRetrievalException {
+	public String grantUserRole(@RequestBody GrantRoleJSONObject grantRoleObj) throws InvalidArgumentsException, UserRetrievalException, UserManagementException, UserPermissionRetrievalException, JsonProcessingException, EntityCreationException, EntityRetrievalException {
 		
-		String isSuccess = String.valueOf(false);
 		UserDTO user = userManager.getByName(grantRoleObj.getSubjectName());
 		if(user == null) {
 			throw new InvalidArgumentsException("No user with name " + grantRoleObj.getSubjectName() + " exists in the system.");
@@ -204,27 +252,27 @@ public class UserManagementController {
 		if(grantRoleObj.getRole().equals("ROLE_ADMIN")) {
 			try {
 				userManager.grantAdmin(user.getSubjectName());
-
-				List<CertificationBodyDTO> acbs = acbManager.getAll();
-				for(CertificationBodyDTO acb : acbs) {
-					acbManager.addPermission(acb, user.getId(), BasePermission.ADMINISTRATION);
-				}
 			} catch(AccessDeniedException adEx) {
 				logger.error("User " + Util.getUsername() + " does not have access to grant ROLE_ADMIN");
+				throw adEx;
 			}
 		} else {
 			userManager.grantRole(user.getSubjectName(), grantRoleObj.getRole());
 		}
 
-		isSuccess = String.valueOf(true);
-		return "{\"roleAdded\" : "+isSuccess+" }";
+		UserDTO updated = userManager.getByName(grantRoleObj.getSubjectName());
+		
+		String activityDescription = "User "+user.getSubjectName()+" was granted role "+grantRoleObj.getRole()+".";
+		activityManager.addActivity(ActivityConcept.ACTIVITY_CONCEPT_USER, user.getId(), activityDescription, user, updated);
+		
+		return "{\"roleAdded\" : true }";
 		
 	}
 	
 	@RequestMapping(value="/revoke_role", method= RequestMethod.POST, 
 			consumes= MediaType.APPLICATION_JSON_VALUE,
 			produces="application/json; charset=utf-8")
-	public String revokeUserRole(@RequestBody GrantRoleJSONObject grantRoleObj) throws InvalidArgumentsException, UserRetrievalException, UserManagementException, UserPermissionRetrievalException {
+	public String revokeUserRole(@RequestBody GrantRoleJSONObject grantRoleObj) throws InvalidArgumentsException, UserRetrievalException, UserManagementException, UserPermissionRetrievalException, JsonProcessingException, EntityCreationException, EntityRetrievalException {
 		
 		String isSuccess = String.valueOf(false);
 		UserDTO user = userManager.getByName(grantRoleObj.getSubjectName());
@@ -235,12 +283,6 @@ public class UserManagementController {
 		if(grantRoleObj.getRole().equals("ROLE_ADMIN")) {
 			try {
 				userManager.removeAdmin(user.getSubjectName());
-				
-				//if they were a chpladmin then they need to have all ACB access removed
-				List<CertificationBodyDTO> acbs = acbManager.getAll();
-				for(CertificationBodyDTO acb : acbs) {
-					acbManager.deletePermission(acb, new PrincipalSid(user.getSubjectName()), BasePermission.ADMINISTRATION);
-				}
 			} catch(AccessDeniedException adEx) {
 				logger.error("User " + Util.getUsername() + " does not have access to revoke ROLE_ADMIN");
 			}
@@ -249,7 +291,7 @@ public class UserManagementController {
 				userManager.removeRole(grantRoleObj.getSubjectName(), grantRoleObj.getRole());
 				
 				//if they were an acb admin then they need to have all ACB access removed
-				List<CertificationBodyDTO> acbs = acbManager.getAll();
+				List<CertificationBodyDTO> acbs = acbManager.getAllForUser();
 				for(CertificationBodyDTO acb : acbs) {
 					acbManager.deletePermission(acb, new PrincipalSid(user.getSubjectName()), BasePermission.ADMINISTRATION);
 				}
@@ -260,7 +302,11 @@ public class UserManagementController {
 			userManager.removeRole(grantRoleObj.getSubjectName(), grantRoleObj.getRole());
 		}
 		
-
+		UserDTO updated = userManager.getByName(grantRoleObj.getSubjectName());
+		
+		String activityDescription = "User "+user.getSubjectName()+" was removed from role "+grantRoleObj.getRole()+".";
+		activityManager.addActivity(ActivityConcept.ACTIVITY_CONCEPT_USER, user.getId(), activityDescription, user, updated);
+		
 		
 		isSuccess = String.valueOf(true);
 		return "{\"roleRemoved\" : "+isSuccess+" }";
@@ -297,46 +343,5 @@ public class UserManagementController {
 		
 		return userManager.getUserInfo(userName);
 		
-	}
-	
-	/**
-	 * create and send the email to invite the user
-	 * @param invitation
-	 */
-	private void sendEmail(InvitationDTO invitation) throws AddressException, MessagingException {
-		 // sets SMTP server properties
-        Properties properties = new Properties();
-        properties.put("mail.smtp.host", "144.202.233.67");
-        properties.put("mail.smtp.port", "25");
-        properties.put("mail.smtp.auth", "true");
-        properties.put("mail.smtp.starttls.enable", "true");
- 
-        // creates a new session with an authenticator
-        javax.mail.Authenticator auth = new javax.mail.Authenticator() {
-            public PasswordAuthentication getPasswordAuthentication() {
-                return new PasswordAuthentication("chpl-etl", "Audac1ous");
-            }
-        };
- 
-        Session session = Session.getInstance(properties, auth);
- 
-        // creates a new e-mail message
-        Message msg = new MimeMessage(session);
- 
-        msg.setFrom(new InternetAddress("chpl-etl@ainq.com"));
-        InternetAddress[] toAddresses = { new InternetAddress(invitation.getEmail()) };
-        msg.setRecipients(Message.RecipientType.TO, toAddresses);
-        msg.setSubject("OpenCHPL Invitation");
-        msg.setSentDate(new Date());
-        // set plain text message
-        msg.setContent(
-        			"<h3>Join OpenCHPL</h3>"
-        			+ "<p>You've been invited to access the CHPL.</p>"
-        			+ "<p>Click the link below to create your account."
-        			+ "<br/>http://localhost:8000/app?hash=" + invitation.getToken() + 
-        			"</p>", "text/html");
-
-        // sends the e-mail
-        Transport.send(msg);
 	}
 }
